Fixed yet another coroutine issue, basically when resuming an if statement with a coroutine the check function would get called again.
To make this happen the atomic steps for looping and for deciding had to be separated.  This is because of the functions
that branch (Goto, Decide, Foreach, etc.) only Goto is "reentrant" i.e. we wouldn't want to resume any of the other atomic
steps when the state is restored.